+++ capa/ida/plugin/form.py
@@ -5,7 +5,6 @@
 # Unless required by applicable law or agreed to in writing, software distributed under the License
 #  is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
-
 import os
 import copy
 import logging
@@ -50,6 +49,7 @@ settings = ida_settings.IDASettings("capa")
 CAPA_SETTINGS_RULE_PATH = "rule_path"
 CAPA_SETTINGS_RULEGEN_AUTHOR = "rulegen_author"
 CAPA_SETTINGS_RULEGEN_SCOPE = "rulegen_scope"
+CAPA_SETTINGS_ANALYZE = "analyze"
 
 
 CAPA_OFFICIAL_RULESET_URL = f"https://github.com/mandiant/capa-rules/releases/tag/v{capa.version.__version__}"
@@ -58,10 +58,20 @@ CAPA_RULESET_DOC_URL = "https://github.com/mandiant/capa/blob/master/doc/rules.m
 
 from enum import IntFlag
 
+OPTION_UNDEFINED = -1
+
 
 class Options(IntFlag):
-    DEFAULT = 0
-    ANALYZE = 1  # Runs the analysis when starting the explorer
+    DEFAULT = 0  # No analysis
+    ANALYZE_AUTO = 1  # Runs the analysis when starting the explorer, see details below
+    ANALYZE_ASK = 2
+
+
+AnalyzeOptionsText = {
+    Options.DEFAULT: "Do not analyze",
+    Options.ANALYZE_AUTO: "Analyze on plugin start (load existing data)",
+    Options.ANALYZE_ASK: "Analyze on plugin start (ask to load or reanalyze)",
+}
 
 
 def write_file(path, data):
@@ -115,6 +125,10 @@ class CapaSettingsInputDialog(QtWidgets.QDialog):
         self.edit_rule_author = QtWidgets.QLineEdit(settings.user.get(CAPA_SETTINGS_RULEGEN_AUTHOR, ""))
         self.edit_rule_scope = QtWidgets.QComboBox()
         self.edit_rules_link = QtWidgets.QLabel()
+        self.edit_analyze = QtWidgets.QComboBox()
+        self.edit_delete_results = QtWidgets.QPushButton(
+            self.style().standardIcon(QtWidgets.QStyle.SP_BrowserStop), "Delete saved capa results"
+        )
 
         self.edit_rules_link.setText(
             f'<a href="{CAPA_OFFICIAL_RULESET_URL}">Download and extract official capa rules</a>'
@@ -122,9 +136,11 @@ class CapaSettingsInputDialog(QtWidgets.QDialog):
         self.edit_rules_link.setOpenExternalLinks(True)
 
         scopes = ("file", "function", "basic block", "instruction")
-
         self.edit_rule_scope.addItems(scopes)
         self.edit_rule_scope.setCurrentIndex(scopes.index(settings.user.get(CAPA_SETTINGS_RULEGEN_SCOPE, "function")))
+
+        self.edit_analyze.addItems(AnalyzeOptionsText.values())
+        self.edit_analyze.setCurrentIndex(settings.user.get(CAPA_SETTINGS_ANALYZE, Options.ANALYZE_AUTO))
 
         buttons = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel, self)
 
@@ -133,6 +149,14 @@ class CapaSettingsInputDialog(QtWidgets.QDialog):
         layout.addRow("", self.edit_rules_link)
         layout.addRow("Default rule author", self.edit_rule_author)
         layout.addRow("Default rule scope", self.edit_rule_scope)
+        layout.addRow("Plugin start option", self.edit_analyze)
+
+        if capa.ida.helpers.idb_contains_capa_results():
+            self.edit_delete_results.clicked.connect(capa.ida.helpers.delete_results)
+            self.edit_delete_results.clicked.connect(lambda state: self.edit_delete_results.setEnabled(False))
+        else:
+            self.edit_delete_results.setEnabled(False)
+        layout.addRow("", self.edit_delete_results)
 
         layout.addWidget(buttons)
 
@@ -141,7 +165,12 @@ class CapaSettingsInputDialog(QtWidgets.QDialog):
 
     def get_values(self):
         """ """
-        return self.edit_rule_path.text(), self.edit_rule_author.text(), self.edit_rule_scope.currentText()
+        return (
+            self.edit_rule_path.text(),
+            self.edit_rule_author.text(),
+            self.edit_rule_scope.currentText(),
+            self.edit_analyze.currentIndex(),
+        )
 
 
 class CapaExplorerForm(idaapi.PluginForm):
@@ -195,8 +224,12 @@ class CapaExplorerForm(idaapi.PluginForm):
 
         self.Show()
 
-        if (option & Options.ANALYZE) == Options.ANALYZE:
-            self.analyze_program()
+        if (
+            settings.user.get(CAPA_SETTINGS_ANALYZE) != Options.DEFAULT
+            or (option & Options.ANALYZE_AUTO) == Options.ANALYZE_AUTO
+        ):
+            analyze = settings.user.get(CAPA_SETTINGS_ANALYZE, OPTION_UNDEFINED)
+            self.analyze_program(analyze=analyze)
 
     def OnCreate(self, form):
         """called when plugin form is created
@@ -252,6 +285,7 @@ class CapaExplorerForm(idaapi.PluginForm):
         self.load_view_tree_tab()
         self.load_view_rulegen_tab()
         self.load_view_status_label()
+        self.load_result_store_label()
         self.load_view_buttons()
 
         # load parent view
@@ -286,10 +320,16 @@ class CapaExplorerForm(idaapi.PluginForm):
 
         self.view_status_label = label
 
+    def load_result_store_label(self):
+        """load result store label"""
+        label = QtWidgets.QLabel()
+        label.setAlignment(QtCore.Qt.AlignLeft)
+        self.view_result_store_label = label
+
     def load_view_buttons(self):
         """load the button controls"""
-        analyze_button = QtWidgets.QPushButton("Analyze")
-        reset_button = QtWidgets.QPushButton("Reset")
+        analyze_button = QtWidgets.QPushButton("Analyze/Load")
+        reset_button = QtWidgets.QPushButton("Reset selections/highlights")
         save_button = QtWidgets.QPushButton("Save")
         settings_button = QtWidgets.QPushButton("Settings")
 
@@ -326,6 +366,7 @@ class CapaExplorerForm(idaapi.PluginForm):
         layout.addWidget(self.view_tabs)
         layout.addLayout(self.view_buttons)
         layout.addWidget(self.view_status_label)
+        layout.addWidget(self.view_result_store_label)
 
         self.parent.setLayout(layout)
 
@@ -611,12 +652,48 @@ class CapaExplorerForm(idaapi.PluginForm):
 
         return None
 
-    def load_capa_results(self, use_cache=False):
+    def load_capa_results(self, use_cache, analyze):
         """run capa analysis and render results in UI
 
         note: this function must always return, exception or not, in order for plugin to safely close the IDA
         wait box
         """
+        if analyze != OPTION_UNDEFINED and analyze != Options.DEFAULT:
+            if capa.ida.helpers.idb_contains_capa_results():
+                if analyze == Options.ANALYZE_AUTO:
+                    load_existing = True
+                elif analyze == Options.ANALYZE_ASK:
+                    self.set_view_result_store_label("")
+                    results: capa.render.result_document.ResultDocument = capa.ida.helpers.load_results()
+                    btn_id = ida_kernwin.ask_buttons(
+                        "Load existing results",
+                        "Reanalyze program",
+                        "",
+                        ida_kernwin.ASKBTN_YES,
+                        f"This database contains capa results generated on "
+                        f"{results.meta.timestamp.strftime('%Y-%m-%d at %H:%M:%S')}.\n"
+                        f"Load existing data or analyze program again?",
+                    )
+                    if btn_id == ida_kernwin.ASKBTN_CANCEL:
+                        self.update_result_store_label()
+                        return True
+                    load_existing = btn_id == ida_kernwin.ASKBTN_YES
+                else:
+                    raise ValueError("missing analysis option %s", analyze)
+
+                if load_existing:
+                    # TODO try/catch
+                    self.resdoc_cache = capa.ida.helpers.load_results()
+                    self.ruleset_cache = capa.ida.helpers.load_ruleset()
+                    self.set_view_result_store_label("loaded existing capa results from database")
+
+                    self.model_data.render_capa_doc(self.resdoc_cache, self.view_show_results_by_function.isChecked())
+                    self.set_view_status_label(
+                        "capa rules: %s (%d rules)"
+                        % (settings.user[CAPA_SETTINGS_RULE_PATH], len(self.ruleset_cache.rules))
+                    )
+                    return True
+
         if not use_cache:
             # new analysis, new doc
             self.resdoc_cache = None
@@ -722,6 +799,17 @@ class CapaExplorerForm(idaapi.PluginForm):
                 logger.error("Failed to collect results (error: %s)", e, exc_info=True)
                 return False
 
+            # persist data across IDA runs
+            try:
+                capa.ida.helpers.save_results(self.resdoc_cache, self.ruleset_cache)
+                self.set_view_result_store_label("saved capa results to database")
+            except Exception as e:
+                self.set_view_result_store_label(
+                    "error saving results to database, see console output for more details."
+                )
+                logger.error("%s", e, exc_info=True)
+                return False
+
         try:
             # either the results are cached and the doc already exists,
             # or the doc was just created above
@@ -749,7 +837,7 @@ class CapaExplorerForm(idaapi.PluginForm):
         self.view_search_bar.setText("")
         self.view_tree.reset_ui()
 
-    def analyze_program(self, use_cache=False):
+    def analyze_program(self, use_cache=False, analyze=Options.ANALYZE_ASK):
         """ """
         self.range_model_proxy.invalidate()
         self.search_model_proxy.invalidate()
@@ -758,7 +846,11 @@ class CapaExplorerForm(idaapi.PluginForm):
         self.set_view_status_label("Loading...")
 
         ida_kernwin.show_wait_box("capa explorer")
-        success = self.load_capa_results(use_cache)
+        try:
+            success = self.load_capa_results(use_cache, analyze)
+        except ValueError as e:
+            success = False
+            logger.error("%s", e, exc_info=True)
         ida_kernwin.hide_wait_box()
 
         self.reset_view_tree()
@@ -1105,6 +1197,7 @@ class CapaExplorerForm(idaapi.PluginForm):
                 settings.user[CAPA_SETTINGS_RULE_PATH],
                 settings.user[CAPA_SETTINGS_RULEGEN_AUTHOR],
                 settings.user[CAPA_SETTINGS_RULEGEN_SCOPE],
+                settings.user[CAPA_SETTINGS_ANALYZE],
             ) = dialog.get_values()
 
     def save_program_analysis(self):
@@ -1216,3 +1309,13 @@ class CapaExplorerForm(idaapi.PluginForm):
         @param text: updated text
         """
         self.view_status_label.setText(text)
+
+    def update_result_store_label(self):
+        if capa.ida.helpers.idb_contains_capa_results():
+            text = "Database contains capa results"
+        else:
+            text = "Database does not contain capa results"
+        self.set_view_result_store_label(text)
+
+    def set_view_result_store_label(self, text):
+        self.view_result_store_label.setText(text)